refactor(builtin-skills): split playwright facade
