+++ src/features/builtin-skills/skills.test.ts
@@ -2,6 +2,7 @@
 
 import { describe, test, expect } from "bun:test"
 import { createBuiltinSkills } from "./skills"
+import { agentBrowserSkill, playwrightSkill } from "./skills/playwright"
 
 describe("createBuiltinSkills", () => {
 	test("returns playwright skill by default", () => {
@@ -15,6 +16,22 @@ describe("createBuiltinSkills", () => {
 		expect(browserSkill).toBeDefined()
 		expect(browserSkill?.description).toContain("browser")
 		expect(browserSkill?.mcpConfig?.playwright).toBeDefined()
+	})
+
+	test("exports browser skill contracts with stable tool surfaces", () => {
+		// #given - direct browser skill exports
+
+		// #when
+		const playwrightMcp = playwrightSkill.mcpConfig?.playwright
+
+		// #then
+		expect(playwrightSkill.name).toBe("playwright")
+		expect(playwrightMcp?.command).toBe("npx")
+		expect(playwrightMcp?.args).toEqual(["@playwright/mcp@latest"])
+		expect(agentBrowserSkill.name).toBe("agent-browser")
+		expect(agentBrowserSkill.allowedTools).toEqual(["Bash(agent-browser:*)"])
+		expect(agentBrowserSkill.template).toContain("agent-browser snapshot -i")
+		expect(agentBrowserSkill.template).toContain("AGENT_BROWSER_SESSION")
 	})
 
 	test("returns playwright skill when browserProvider is 'playwright'", () => {
